libxl: ocaml: switch all functions over to take a context
# HG changeset patch
# User Rob Hoes <rob.hoes@citrix.com>
# Date 1359496878 0
# Node ID 2283147b584ac35dccfe5f84976226ca237c5184
# Parent  7315513a8421f7d19ab81a95468e7e2b0adf0dea
libxl: ocaml: switch all functions over to take a context.

Since the context has a logger we can get rid of the logger built into these
bindings and use the xentoollog bindings instead.

The gc is of limited use when most things are freed with libxl_FOO_dispose,
so get rid of that too.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>